Fix propagations of quantities through procurements. (#86)
* Fix propagations of quantities through procurements.

Instead of modifying quantities on stock moves, we run the procurements with negative quantities.
This is supported out of the box by the procurement.group.run method.

If stock moves must be cancelled and unlinked from their pickings, we do so recursively.
Otherwise, preparation moves are not cancelled and stay in a confirmed state.