Simplify project_iteration (#96)
* Simplify project_iteration

Use a many2one field to store the relation between task and parent project
instead of a many2many relationship.

The many2many was overkill and prevented from grouping tasks by parent project.

Also fix a bug. If a task's project has no parent and is not a parent project,
then the parent project for the task must be undefined.

* Bump module versions