TA#29141 refactor project_time_range
Previously, 5 modules were available:

* project_budget_management
* project_task_subtask_time_range
* project_task_time_range
* project_time_management
* project_time_range

This number of modules does not represent well the number of functionnal requirements
related to these modules.

There are 2 different requirements:

* Estimating tasks and have visualize the estimations for a whole project
* Budgetize the time before a project is started

The new module project_time_range adds everyting needed on tasks, subtasks and projects
in regard to the first requirement.

The new module project_time_budget adds extra fields on projects in the context of a budgetized project.