TA#milestone_no_ref
bug correction:
when a project is copied, we need change milestone reference for
deactivated tasks too

bug correction :
when a milestone has his project modified, all his associated tasks not
associated to this new projector no project are dissociated.

Add the field "Active" boolean on milestones and a button is displayed
on form view (as in projects and tasks, necessary for future module
sale_project_milestone)

Add field "Active_toggle" boolean which has the value of field active
when button Active in form view is checked.

when a project change field "Use milestones", milestones which have
field a"Active_toggle" set to True are set to same value.

When a project is (de)activated, milestones, which have field
"Active_toggle" set to True, too.